moved character movement into Home.js for game development
